+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added v4 Categories beta APIs to v4 client
 - Added v4 VolumeGroups beta APIs to v4 client
+- Added an adapter client for taking care of v3-v4 client compatibility
 
 ### Changed
 - Updated v4 API clients from v4 alpha to v4 beta  APIs